fix(ticket table)
-removed ticket table from admin dashboard
- added ticket table to tickets
